+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - #73: Add new application metric: `circuit_maintenance_status` to show the circuit status depending on related Circuit Maintenances.
+- #78: Add optional `extra_scopes` config parameter to use with Gmail notification sources.
 
 ### Fixed